requirements.txt
aiogram==3.2.0
aiohttp
python-dotenv
